+++ examples/form-example/app/components/form-example.js
@@ -1,26 +1,89 @@
 import React from 'react';
 import yup from 'yup';
 import SchemaBranchMixin from 'baobab-react-schemabranchmixin';
-import {Form, ValidationBox, Input, Radio, CheckBox, Submit} from 'react-form-tools';
+import {
+    Form,
+    ValidationBox,
+    Input,
+    Radio,
+    MultipleCheckBox,
+    CheckBox,
+    Submit
+} from 'react-form-tools';
+
+const NestedForm = React.createClass({
+    displayName: 'NestedForm',
+
+    mixins: [SchemaBranchMixin],
+
+    schema: {
+        form: {
+        },
+    },
+
+    validationSchema: yup.object().shape({}),
+
+    onSubmit(data) {
+        console.log('NestedForm is submitted with valid data', data);
+    },
+
+    onInvalidSubmit(errors) {
+        console.log('NestedForm is submitted with errors', errors);
+    },
+
+    render() {
+        return (
+            <Form
+                style={{ border: '1px solid black', padding: 5, marginTop: 5, marginBottom: 5 }}
+                cursor={this.cursors.form}
+                validationSchema={this.validationSchema}
+                onSubmit={this.onSubmit}
+                onInvalidSubmit={this.onInvalidSubmit}>
+                Nested form
+                <div>
+                    <span>Text Input</span>
+                    <ValidationBox fieldPath="fieldTextInput">
+                        <Input />
+                    </ValidationBox>
+                </div>
+
+                <div>
+                    <Submit>Submit</Submit>
+                </div>
+            </Form>
+        );
+    },
+});
 
 export default React.createClass({
     mixins: [SchemaBranchMixin],
 
     schema: {
-        form: {},
+        nestedForm: {},
+        form: {
+            fieldMultipleCheckBoxInput: [],
+        },
         formState: {},
     },
 
-    validationSchema: yup.object().shape({
+    validationSchema: yup.object().shape({}),
 
-    }),
+    onSubmit(data) {
+        console.log('Form is submitted with valid data', data);
+    },
+
+    onInvalidSubmit(errors) {
+        console.log('Form is submitted with errors', errors);
+    },
 
     render() {
         return (
             <Form
                 cursor={this.cursors.form}
                 formStateCursor={this.cursors.formState}
-                validationSchema={this.validationSchema}>
+                validationSchema={this.validationSchema}
+                onSubmit={this.onSubmit}
+                onInvalidSubmit={this.onInvalidSubmit}>
                 <div>
                     <span>Text Input</span>
                     <ValidationBox fieldPath="fieldTextInput">
@@ -50,10 +113,32 @@ export default React.createClass({
                 </div>
 
                 <div>
+                    <ValidationBox fieldPath="fieldMultipleCheckBoxInput">
+                        <label>
+                            <MultipleCheckBox value="1" />
+                            1
+                        </label>
+                        <label>
+                            <MultipleCheckBox value="2" />
+                            2
+                        </label>
+                        <label>
+                            <MultipleCheckBox value="3" />
+                            3
+                        </label>
+                    </ValidationBox>
+                </div>
+
+                <div>
                     <ValidationBox fieldPath="fieldCheckBoxInput">
                         <CheckBox value={true} /> Agree
                     </ValidationBox>
                 </div>
+
+                <div>
+                    <NestedForm tree={this.cursors.nestedForm} />
+                </div>
+
                 <div>
                     <Submit>Submit</Submit>
                 </div>