+++ src/components/input.js
@@ -3,16 +3,19 @@ import ReactDOM from 'react-dom';
 import _ from 'lodash';
 import PureRenderMixin from 'react-addons-pure-render-mixin';
 import BaobabPropTypes from 'baobab-prop-types';
+import { BranchMixin } from 'baobab-react-mixins';
 import { FormComponentMixin } from '../mixins';
 
 export default React.createClass({
     displayName: 'Input',
 
-    mixins: [PureRenderMixin, FormComponentMixin],
+    mixins: [BranchMixin, FormComponentMixin, PureRenderMixin],
 
     propTypes: {
         cursor: BaobabPropTypes.cursor,
         onChange: React.PropTypes.func,
+        onKeyPress: React.PropTypes.func,
+        onSync: React.PropTypes.func,
         onBlur: React.PropTypes.func,
         sync: React.PropTypes.bool,
         syncOnlyOnBlur: React.PropTypes.bool,
@@ -24,6 +27,12 @@ export default React.createClass({
     deferredSyncTimer: null,
 
     msToPoll: 200,
+
+    cursors(props, context) {
+        return {
+            cursorValue: this.getCursor(props, context),
+        };
+    },
 
     getInitialState() {
         return {
@@ -40,24 +49,33 @@ export default React.createClass({
             onBlur: _.identity,
             onChange: _.identity,
             onSync: _.identity,
+            onKeyPress: _.identity,
             sync: false,
         };
     },
 
-    componentWillReceiveProps(nextProps, nextContext) {
+    componentDidUpdate(prevProps, prevState, prevContext) {
         if (this.deferredSyncTimer) {
             // Does not set state when component received props while user inputs
             return;
         }
 
-        this.setState({
-            value: this.props.toInternal(this.getCursor(nextProps, nextContext).get()),
-        });
+        if (this.state.cursorValue !== prevState.cursorValue) {
+            this.setState({
+                value: this.state.cursorValue,
+            });
+        }
     },
 
     componentDidMount() {
         if (this.props.autoFocus) {
-            setTimeout(() => ReactDOM.findDOMNode(this.refs.input).focus(), 0);
+            setTimeout(() => {
+                const element = ReactDOM.findDOMNode(this.refs.input);
+
+                if (element) {
+                    element.focus();
+                }
+            }, 0);
         }
     },
 
@@ -72,62 +90,75 @@ export default React.createClass({
         }
     },
 
-    deferredSyncValue() {
+    deferredSyncValue(eventCallback) {
         this.clearDeferredSyncTimer();
-        this.deferredSyncTimer = setTimeout(this.syncValue, this.msToPoll);
+        this.deferredSyncTimer = setTimeout(() => this.syncValue(eventCallback), this.msToPoll);
     },
 
-    syncValue() {
+    syncValue(eventCallback) {
+        // Synchronizes value with cursor
         const value = this.props.nullable && this.state.value === '' ? null : this.state.value;
-        const previousValue = this.getCursor().get();
+        const previousValue = this.state.cursorValue;
 
         if (value === previousValue) {
+            if (_.isFunction(eventCallback)) {
+                eventCallback();
+            }
+
             return;
         }
 
-        this.getCursor().set(value);
-        this.setDirtyState();
+        this.setValue(value, () => {
+            this.setDirtyState();
+            this.props.onSync(value, previousValue);
 
-        // Wait for next frame
-        setTimeout(() => this.props.onSync(value, previousValue), 0);
-    },
-
-    setValue(value, forceSync=false) {
-        this.setState({ value }, function () {
-            if (this.props.sync || forceSync) {
-                this.syncValue();
-                return;
-            }
-
-            if (!this.props.syncOnlyOnBlur) {
-                this.deferredSyncValue();
+            if (_.isFunction(eventCallback)) {
+                eventCallback();
             }
         });
     },
 
-    onChange(evt) {
-        const value = this.props.toInternal(evt.target.value);
+    updateValue(value, forceSync=false, eventCallback) {
+        // Synchronizes value with state
+        this.setState({ value }, function () {
+            if (this.props.sync || forceSync) {
+                this.syncValue(eventCallback);
+                return;
+            }
+
+            if (!this.props.syncOnlyOnBlur) {
+                this.deferredSyncValue(eventCallback);
+            }
+        });
+    },
+
+    onChange(event) {
+        const value = this.props.toInternal(event.target.value);
         const previousValue = this.state.value;
 
         if (value === previousValue) {
             return;
         }
 
-        this.setValue(value);
+        this.updateValue(value);
         this.props.onChange(value, previousValue);
     },
 
-    onBlur(evt) {
-        const value = this.props.toInternal(evt.target.value);
+    onBlur(event) {
+        const value = this.props.toInternal(event.target.value);
 
         // Prevent future updates
         this.clearDeferredSyncTimer();
 
         // Set inner state value and force synchronization
-        this.setValue(value, true);
+        this.updateValue(value, true, () => this.props.onBlur(event));
+    },
 
-        // Wait for next frame
-        setTimeout(() => this.props.onBlur(evt), 0);
+    onKeyPress(event) {
+        this.processKeyPress(event, () => {
+            this.clearDeferredSyncTimer();
+            this.syncValue(() => this.context.form.submit());
+        });
     },
 
     render() {
@@ -144,7 +175,10 @@ export default React.createClass({
         }
 
         return (
-            <input type={this.props.type} {...this.props} {...props} ref="input" />
+            <input {...this.props} {...props}
+                type={this.props.type}
+                onKeyPress={this.onKeyPress}
+                ref="input" />
         );
     },
 });