+++ src/mixins.js
@@ -1,4 +1,6 @@
 import React from 'react';
+import _ from 'lodash';
+import { isEnterPressed, getFieldPathAsArray } from './utils';
 
 export const FormComponentMixin = {
     contextTypes: {
@@ -6,21 +8,60 @@ export const FormComponentMixin = {
         fieldPath: React.PropTypes.array,
     },
 
+    processKeyPressForSubmit(event) {
+        // Helper method for form components
+        // Submits form on enter by default
+        this.processKeyPress(event, () => this.context.form.submit());
+    },
+
+    processKeyPress(event, fn) {
+        // Callback `fn` will be called on enter press
+        if (!this.context.form.isHtmlForm()) {
+            if (isEnterPressed(event)) {
+                event.preventDefault();
+                event.stopPropagation();
+                fn();
+            }
+        }
+
+        if (_.isFunction(this.props.onKeyPress)) {
+            this.props.onKeyPress(event);
+        }
+    },
+
     getCursor(props, context) {
         props = props || this.props;
         context = context || this.context;
 
-        /* istanbul ignore next */
-        if (!props.cursor && !context.fieldPath) {
-            throw 'react-form.tools ' + this.displayName + ': cursor must be set via `cursor` or ' +
-                  'via higher order component ValidationBox with fieldPath';
+        if (props.cursor) {
+            return props.cursor;
         }
 
-        return props.cursor || context.form.cursor.select(context.fieldPath);
+        if (props.fieldPath) {
+            return context.form.cursor.select(getFieldPathAsArray(props.fieldPath));
+        }
+
+        if (context.fieldPath) {
+            return context.form.cursor.select(context.fieldPath);
+        }
+
+        /* istanbul ignore next */
+        throw `react-form.tools ${this.displayName}: cursor must be set via 'cursor',
+               'fieldPath' or via higher order component ValidationBox with 'fieldPath'`;
     },
 
     inValidationBox() {
         return !!(this.context.form && this.context.fieldPath);
+    },
+
+    setValue(value, callback) {
+        const cursor = this.getCursor();
+
+        if (_.isFunction(callback)) {
+            cursor.once('update', callback);
+        }
+
+        cursor.set(value);
     },
 
     setDirtyState() {