+++ test/components/validation-error.spec.js
@@ -5,8 +5,8 @@ import Baobab from 'baobab';
 import SchemaBranchMixin from 'baobab-react-schemabranchmixin';
 import TestUtils from 'react-addons-test-utils';
 import yup from 'yup';
-import {Form, ValidationError} from '../../src/components';
-import {Root} from '../utils';
+import { Form, ValidationError } from '../../src/components';
+import { Root } from '../utils';
 
 const tree = new Baobab(
     {},
@@ -31,8 +31,10 @@ const FormWithValidationError = React.createClass({
 
     render() {
         return (
-            <Form cursor={this.cursors.form} validationSchema={this.validationSchema} ref="form">
-                <ValidationError fieldPath="field" ref="error" {...this.props.errorProps} />
+            <Form cursor={this.cursors.form}
+                validationSchema={this.validationSchema} ref="form">
+                <ValidationError fieldPath="field"
+                    ref="error" {...this.props.errorProps} />
             </Form>
         );
     },